requirements.txt
flask
flask-socketio
websocket-client==1.6.4
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
plotly==5.17.0
dash==2.13.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
loguru==0.7.2